Rename: slot 실패하고, 원상복귀
