--- src/components/generation/PassageLayout.vue
@@ -1,29 +1,0 @@
-<template>
-    <div class="app-container">
-        <p id="main-title">지문 생성</p>
-        <div class="main-content">
-            <slot></slot>
-        </div>
-    </div>
-</template>
-<script setup>
-
-</script>
-<style scoped>
-#main-title {
-    position: absolute;
-    width: 87px;
-    height: 36px;
-    left: 292px;
-    top: 70px;
-
-    font-family: 'Pretendard';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 150%;
-
-    letter-spacing: -0.02em;
-    color: #000000;
-}
-</style>
