+++ src/router/generation.routes.js
@@ -1,4 +1,4 @@
-import GenerationMain from "@/components/generation/passage/PassageMain.vue";
+import GenerationMain from "@/components/generation/PassageMain.vue";
 import { createRouter, createWebHistory } from "vue-router";
 
 const router = createRouter({
